+++ packaging/crosswalk-include-tizen-ime-files.patch
@@ -43,10 +43,10 @@ index 785e977..e473043 100644
      }],
 +    ['tizen_mobile == 1', {
 +      'sources': [
-+        '<(DEPTH)/xwalk/ime/tizen-scim/input_method_scim.cc',
-+        '<(DEPTH)/xwalk/ime/tizen-scim/input_method_scim.h',
-+        '<(DEPTH)/xwalk/ime/tizen-scim/scim_bridge.cc',
-+        '<(DEPTH)/xwalk/ime/tizen-scim/scim_bridge.h',
++        '<(DEPTH)/xwalk/tizen/ime/input_method_scim.cc',
++        '<(DEPTH)/xwalk/tizen/ime/input_method_scim.h',
++        '<(DEPTH)/xwalk/tizen/ime/scim_bridge.cc',
++        '<(DEPTH)/xwalk/tizen/ime/scim_bridge.h',
 +      ],
 +      'dependencies': [
 +        '../build/linux/system.gyp:scim',
@@ -66,7 +66,7 @@ index 03bd5e4..06e8cff 100644
  #include "ui/base/ime/input_method_imm32.h"
  #include "ui/base/ime/input_method_tsf.h"
 +#elif defined(OS_TIZEN_MOBILE)
-+#include "xwalk/ime/tizen-scim/input_method_scim.h"
++#include "xwalk/tizen/ime/input_method_scim.h"
  #elif defined(USE_AURA) && defined(USE_X11)
  #include "ui/base/ime/input_method_linux_x11.h"
  #else
