+++ tizen/ime/README
