+++ tizen/ime/scim_bridge.cc
